230206_디버깅

+++ components/mos/include/_kf.h
@@ -28,10 +28,10 @@ typedef struct {
 }kf_ars_t;
 
 float SystemModel_ars(float q[4], float Omega[3]) ;
-void qut_discrete(int32_t gyro[3]);
-void Kalman_Gain(int32_t acc_lpf[3]);
-void init_kf_ars(int32_t acc[3]);
-void cov_discrete(int32_t gyro[3]);
-float* kf_ars_loop(int32_t acc_lpf[3], int32_t gyro[3]);
-void Measurement_Update(int32_t acc_lpf[3]);
+void qut_discrete(float gyro[3]);
+void Kalman_Gain(float acc_lpf[3]);
+void init_kf_ars(float acc[3]);
+void cov_discrete(float gyro[3]);
+float* kf_ars_loop(float acc_lpf[3], float gyro[3]);
+void Measurement_Update(float acc_lpf[3]);
 #endif
